show purchases in table

+++ src/vendor_app/app/src/main/kotlin/cz/quanti/android/vendor_app/repository/purchase/dto/TransactionPurchase.kt
@@ -2,7 +2,7 @@ package cz.quanti.android.vendor_app.repository.purchase.dto
 
 class TransactionPurchase (
     var purchaseId: Long = 0,
-    var value: Long = 0,
+    var value: Double = 0.0,
     var currency: String = "",
     var beneficiaryId: Long = 0,
     var createdAt: String = "",
