+++ src/vendor_app/app/src/main/kotlin/cz/quanti/android/vendor_app/repository/purchase/dto/api/V2TransactionApiEntity.kt
@@ -6,8 +6,8 @@ data class V2TransactionsApiEntity (
 )
 
 data class TransactionApiEntity(
-    var projectId: Int = 0,
+    var projectId: Long = 0,
     var purchaseIds: List<Int> = listOf(),
-    var value: Long = 0,
+    var value: Double = 0.0,
     var currency: String = ""
 )
